requirements.txt
torch==2.9.0
torchvision==0.24.0
timm==1.0.21
torchmetrics==1.8.2
numpy==2.2.6
scikit-learn==1.7.2
pandas==2.3.3
matplotlib==3.10.7
opencv-python==4.12.0.88
Pillow==12.0.0
tqdm==4.67.1
PyYAML==6.0.3
wandb==0.18.5
pytest==8.3.3
black==24.10.0
ruff==0.7.1
isort==5.13.2
